Remove rescue block from the checkout advance service, move ErrorHandler from spree/api to spree/core
