+++ api/lib/spree/api/dependencies.rb
@@ -86,7 +86,7 @@ module Spree
         storefront_taxon_finder: -> { Spree::Dependencies.taxon_finder },
         storefront_variant_finder: -> { Spree::Dependencies.variant_finder },
 
-        error_handler: 'Spree::Api::ErrorHandler',
+        error_handler: 'Spree::ErrorHandler',
 
         # serializers
         platform_admin_user_serializer: 'Spree::Api::V2::Platform::AdminUserSerializer',
