+++ api/spec/controllers/spree/api/v2/base_controller_spec.rb
@@ -110,7 +110,7 @@ describe Spree::Api::V2::BaseController, type: :controller do
       it do
         expect(subject).to receive(:index).and_raise(exception)
         expect(subject).to receive(:spree_current_user).at_least(:once).and_return(user)
-        expect_next_instance_of(::Spree::Api::ErrorHandler) do |instance|
+        expect_next_instance_of(::Spree::ErrorHandler) do |instance|
           expect(instance).to receive(:call).with(
             exception: exception,
             opts: { user: user }
